change status datatype

+++ FinancialPlanningBE/FinancialPlanning.Data/Repositories/ReportRepository.cs
@@ -140,7 +140,7 @@ namespace FinancialPlanning.Data.Repositories
         {
             foreach (var report in reports)
             {
-                report.Status = (int)Common.ReportStatus.Closed;
+                report.Status = Common.ReportStatus.Closed;
             }
 
             await _context.SaveChangesAsync();
